Address review comments pt. 2
