+++ src/core/file_sys/fs_path_utility.h
@@ -91,12 +91,8 @@ public:
     }
 
 #define DECLARE_PATH_FLAG_HANDLER(__WHICH__)                                                       \
-    constexpr bool Is##__WHICH__##Allowed() const {                                                \
-        return (m_value & __WHICH__##Flag) != 0;                                                   \
-    }                                                                                              \
-    constexpr void Allow##__WHICH__() {                                                            \
-        m_value |= __WHICH__##Flag;                                                                \
-    }
+    constexpr bool Is##__WHICH__##Allowed() const { return (m_value & __WHICH__##Flag) != 0; }     \
+    constexpr void Allow##__WHICH__() { m_value |= __WHICH__##Flag; }
 
     DECLARE_PATH_FLAG_HANDLER(WindowsPath)
     DECLARE_PATH_FLAG_HANDLER(RelativePath)
